folded up, yo

+++ Take 2/Assets/Scripts/BulletController.cs
@@ -3,8 +3,8 @@ using System.Collections;
 
 public class BulletController : MonoBehaviour {
 
-    private float damage;
-    public float Damage
+    private uint damage;
+    public uint Damage
     {
         set { damage = value; }
         get { return damage; }
